Unify usage of ThriftMetastoreAuthenticationType
